requirements.txt
streamlit
openai
google-generativeai
huggingface_hub
requests
python-dotenv